+++ src/clojure/lazytest/find.clj
@@ -2,11 +2,11 @@
   (:require
    [lazytest.core :refer [describe it]]
    [lazytest.malli]
-   [lazytest.suite :refer [suite suite? test-seq]]
+   [lazytest.suite :refer [suite suite?]]
    [lazytest.test-case :refer [test-case?]]))
 
 (defn- set-var [value this-var]
-  (vary-meta value assoc :type :lazytest/test-var :var this-var))
+  (assoc value :type :lazytest/var :var this-var))
 
 (defn find-var-test-value
   [this-var]
@@ -16,21 +16,48 @@
           test-metadata (:test m)]
       (cond
         ;; (defdescribe example ...)
-        ;; (def example (suite ...))
+        (= :lazytest/var (:type m))
+        (set-var (value) this-var)
+        ;; (def example (describe ...))
         (suite? value)
         (set-var value this-var)
         ;; (defn example {:test (describe ...)})
-        ;; (defn example {:test (suite ...)})
         (suite? test-metadata)
-        (set-var (describe this-var test-metadata) this-var)
+        (set-var test-metadata this-var)
         ;; (defn example {:test (it ...)})
         (test-case? test-metadata)
-        (set-var (suite (test-seq [test-metadata])) this-var)
+        (let [new-test (describe this-var)]
+          (set-var (update new-test :tests conj test-metadata) this-var))
         ;; (defn example {:test #(expect ...)})
         (fn? test-metadata)
-        (set-var (suite (test-seq [(it "`:test` metadata" (test-metadata))])) this-var)))))
+        (set-var (describe this-var (it "`:test` metadata" (test-metadata))) this-var)))))
 
-(defn- test-seq-for-ns [this-ns]
+#_(comment
+  (defn test-fn
+    {:test #(expect (= 0 (test-fn 1)))}
+    [a]
+    (+ a a))
+
+  (defn test-test-case
+    {:test (it "test case example"
+             (expect (= 1 (test-test-case 1))))}
+    [a]
+    (+ a a))
+
+  (defn test-describe
+    {:test (describe "top level"
+             (it "test-describe example" (expect (= 1 (test-describe 1))))
+             (it "test-describe example two" (expect (= 0 (test-describe 1)))))}
+    [a]
+    (+ a a))
+
+  (def test-describe-def
+    (describe "test-def-describe"
+      (it "test-def-describe example")
+      (it "test-def-describe example two")))
+  )
+
+(defn- test-suites-for-ns [this-ns]
   (->> (ns-interns this-ns)
        (vals)
        (sort-by (comp (juxt :line :column) meta))
@@ -44,17 +71,23 @@
 
   By default, recurses on all Vars in a namespace looking for values
   for which lazytest.suite/suite? is true. If a namesapce
-  has :test-suite metadata, uses that value instead.
+  has :lazytest/suite metadata, uses that value instead.
 
   Always returns nil for the clojure.core namespace, to avoid special
   Vars such as *1, *2, *3"
   [n]
   (when-not (= (the-ns 'clojure.core) n)
-    (or (:test-suite (meta n))
-        (when-let [s (test-seq-for-ns n)]
-          (vary-meta
-            (suite (test-seq s))
-            merge (meta n) {:type :lazytest/ns-suite :ns-name (ns-name n)})))))
+    (or (:test-suite (meta n)) ;; deprecated, undocumented
+        (:lazytest/suite (meta n))
+        (when-let [s (test-suites-for-ns n)]
+          (-> (meta n)
+              (assoc :suites s)
+              (assoc :doc (str (ns-name n)))
+              (assoc :type :lazytest/ns)
+              (suite))))))
+
+(comment
+  (find-ns-suite *ns*))
 
 (defn find-suite
   "Returns test suite containing suites for the given namespaces.
@@ -62,5 +95,10 @@
   [& names]
   (let [names (or (seq names) (all-ns))
         nses (mapv the-ns names)]
-    (vary-meta (suite (keep find-ns-suite nses))
-               assoc :type :lazytest/run :nses nses)))
+    (suite {:type :lazytest/run
+            :nses nses
+            :suites (keep find-ns-suite nses)})))
+
+(comment
+  (:suites (find-suite)))
+