+++ src/clojure/lazytest/runner.clj
@@ -6,115 +6,108 @@
    [lazytest.malli]
    [lazytest.reporters :as r :refer [nested report]]
    [lazytest.suite :as s :refer [suite-result]]
-   [lazytest.test-case :refer [try-test-case]]
-   [malli.experimental :as mx]))
+   [lazytest.test-case :refer [try-test-case]]))
 
 (set! *warn-on-reflection* true)
 
-(defn dispatch [_config m]
-  (or (:type m) (-> m meta :type)))
+(defn dispatch [m _config] (:type m))
 
-(defmulti run-test #'dispatch)
-(defmethod run-test :default [_config m]
+(defmulti run-tree {:arglists '([m config])} #'dispatch)
+(defmethod run-tree :default [m _config]
   (throw (ex-info "Non-test given to run-suite." {:obj m})))
-(defmethod run-test nil [_config _])
+(defmethod run-tree nil [_m _config])
 
-(defn ->suite-result [config s source-type]
-  (let [sm (meta s)
-        id (s/identifier sm)
+(defn ->suite-result [suite config source-type]
+  (let [id (:doc suite)
         start (System/nanoTime)
         config (-> config
-                    (update ::depth #(if id (inc %) %))
-                    (update ::suite-history conj sm))
-        f (if-let [around-fn (combine-arounds sm)]
+                   (update :lazytest.runner/depth #(if id (inc %) %))
+                   (update :lazytest.runner/suite-history conj suite))
+        f (if-let [around-fn (combine-arounds suite)]
             #(let [ret (volatile! nil)
-                   tests (propagate-eachs sm %)]
-               (around-fn (fn [] (vreset! ret (run-test config tests))))
+                   tests (propagate-eachs suite %)]
+               (around-fn (fn [] (vreset! ret (run-tree tests config))))
                @ret)
-            #(run-test config (propagate-eachs sm %)))
-        results (vec (keep f s))
+            #(let [child (propagate-eachs suite %)]
+               (run-tree child config)))
+        results (vec (keep f (concat (:tests suite)
+                                     (:suites suite))))
         duration (double (- (System/nanoTime) start))]
-    (-> (suite-result s results)
-        (assoc ::source-type source-type)
-        (assoc ::duration duration))))
+    (-> (suite-result suite results)
+        (assoc :lazytest.runner/source-type source-type)
+        (assoc :lazytest.runner/duration duration))))
 
-(defmethod run-test :lazytest/run [config s]
-  (let [sm (meta s)]
-    (report config (assoc sm :type :begin-test-run))
-    (run-befores sm)
-    (let [results (->suite-result config s :lazytest/run)]
-      (report config (assoc sm :type :end-test-run :results results))
-      (run-afters sm)
-      results)))
+(defmethod run-tree :lazytest/run
+  run-test--lazytest-run
+  [suite config]
+  (report config (assoc suite :type :begin-test-run))
+  (run-befores suite)
+  (let [results (->suite-result suite config :lazytest/run)]
+    (report config (assoc suite :type :end-test-run :results results))
+    (run-afters suite)
+    results))
 
-(defmethod run-test :lazytest/ns-suite [config s]
-  (let [sm (meta s)]
-    (report config (assoc sm :type :begin-ns-suite))
-    (run-befores sm)
-    (let [results (->suite-result config s :lazytest/ns-suite)]
-      (report config (assoc sm :type :end-ns-suite :results results))
-      (run-afters sm)
-      results)))
+(defmethod run-tree :lazytest/ns
+  run-test--lazytest-ns
+  [suite config]
+  (report config (assoc suite :type :begin-test-ns))
+  (run-befores suite)
+  (let [results (->suite-result suite config :lazytest/ns)]
+    (report config (assoc suite :type :end-test-ns :results results))
+    (run-afters suite)
+    results))
 
-(defmethod run-test :lazytest/test-var [config s]
-  (let [sm (meta s)]
-    (report config (assoc sm :type :begin-test-var))
-    (run-befores sm)
-    (let [results (->suite-result config s :lazytest/test-var)]
-      (report config (assoc sm :type :end-test-var :results results))
-      (run-afters sm)
-      results)))
+(defmethod run-tree :lazytest/var
+  run-test--lazytest-var
+  [suite config]
+  (report config (assoc suite :type :begin-test-var))
+  (run-befores suite)
+  (let [results (->suite-result suite config :lazytest/var)]
+    (report config (assoc suite :type :end-test-var :results results))
+    (run-afters suite)
+    results))
 
-(defmethod run-test :lazytest/suite [config s]
-  (let [sm (meta s)]
-    (report config (assoc sm :type :begin-test-suite))
-    (run-befores sm)
-    (let [results (->suite-result config s :lazytest/suite)]
-      (report config (assoc sm :type :end-test-suite :results results))
-      (run-afters sm)
-      results)))
+(defmethod run-tree :lazytest/suite
+  run-test--lazytest-suite
+  [suite config]
+  (report config (assoc suite :type :begin-test-suite))
+  (run-befores suite)
+  (let [results (->suite-result suite config :lazytest/suite)]
+    (report config (assoc suite :type :end-test-suite :results results))
+    (run-afters suite)
+    results))
 
-(defmethod run-test :lazytest/test-seq [config s]
-  (let [sm (meta s)]
-    (report config (assoc sm :type :begin-test-seq))
-    (run-befores sm)
-    (let [results (->suite-result config s :lazytest/test-seq)]
-      (report config (assoc sm :type :end-test-seq :results results))
-      (run-afters sm)
-      results)))
-
-(defmethod run-test :lazytest/test-case [config tc]
-  (let [tc-meta (meta tc)
-        start (System/nanoTime)]
-    (report config (assoc tc-meta :type :begin-test-case))
-    (run-befores tc-meta)
-    (let [results (if-let [around-fn (combine-arounds tc-meta)]
+(defmethod run-tree :lazytest/test-case
+  run-suite--lazytest-test-case
+  [tc config]
+  (let [start (System/nanoTime)]
+    (report config (assoc tc :type :begin-test-case))
+    (run-befores tc)
+    (let [f (with-meta (:body tc) tc)
+          results (if-let [around-fn (combine-arounds tc)]
                     (let [ret (volatile! nil)]
-                      (run-before-eachs tc-meta)
-                      (around-fn (fn [] (vreset! ret (try-test-case tc))))
-                      (run-after-eachs tc-meta)
+                      (run-before-eachs tc)
+                      (around-fn (fn [] (vreset! ret (try-test-case f))))
+                      (run-after-eachs tc)
                       @ret)
-                    (do (run-before-eachs tc-meta)
-                        (let [ret (try-test-case tc)]
-                          (run-after-eachs tc-meta)
+                    (do (run-before-eachs tc)
+                        (let [ret (try-test-case f)]
+                          (run-after-eachs tc)
                           ret)))
           duration (double (- (System/nanoTime) start))
           results (assoc results ::duration duration)]
       (report config results)
-      (report config (assoc tc-meta :type :end-test-case :results results))
-      (run-afters tc-meta)
+      (report config (assoc tc :type :end-test-case :results results))
+      (run-afters tc)
       results)))
 
 (defn run-tests
   "Runs tests defined in the given namespaces."
   ([namespaces] (run-tests {:reporter nested} namespaces))
   ([config namespaces]
-   (let [ste (apply find-suite namespaces)
-         tree (filter-tree config ste)
-         result (run-test config tree)]
-     (if (:focus (meta tree))
-       (vary-meta result assoc :focus true)
-       result))))
+   (let [suite (apply find-suite namespaces)
+         suite (filter-tree suite config)]
+     (run-tree suite config))))
 
 (defn run-all-tests
   "Run tests defined in all namespaces."
@@ -122,9 +115,9 @@
   ([config]
    (run-tests config nil)))
 
-(mx/defn run-test-var
-  [config v :- :lt/var]
-  (let [tree (-> (find-var-test-value v)
-                 (vary-meta assoc :type :lazytest/run)
-                 (#(filter-tree config %)))]
-    (run-test config tree)))
+(defn run-test-var
+  [v config]
+  (-> (find-var-test-value v)
+      (assoc :type :lazytest/run)
+      (filter-tree config)
+      (run-tree config)))