+++ src/clojure/lazytest/test_case.clj
@@ -1,29 +1,23 @@
-(ns lazytest.test-case
-  (:require
-    [lazytest.malli]
-    [malli.experimental :as mx]))
+(ns lazytest.test-case)
 
-(defmacro test-case
-  "Sets metadata on function f identifying it as a test case. A test
-  case function may execute arbitrary code and may have side effects.
+^:clj-reload/keep
+(defrecord TestCase [type doc body context ns file line metadata])
+
+(defn test-case? [obj]
+  (instance? TestCase obj))
+
+(defn test-case
+  "A test case function may execute arbitrary code and may have side effects.
   It should throw an exception to indicate failure. Returning without
   throwing an exception indicates success.
 
   Additional identifying metadata may be placed on the function, such
   as :ns-name and :doc."
-  [f]
-  `(let [f# ~f
-         fm# (meta f#)]
-     (vary-meta f# assoc
-                :type :lazytest/test-case
-                :file ~*file*
-                :line (or (:line fm#) ~(:line (meta &form)))
-                ::test-case true)))
-
-(defn test-case?
-  "True if x is a test case."
-  [x]
-  (and (fn? x) (::test-case (meta x))))
+  {:arglists '([{:keys [type doc body context ns file line metadata] :as test-case}])}
+  [base]
+  (map->TestCase (-> base
+                     (update :context #(or % {}))
+                     (assoc :type :lazytest/test-case))))
 
 (defn- stacktrace-file-and-line
   "Adapted from clojure.test"
@@ -42,7 +36,7 @@
         m (merge source-meta thrown-data caught-data)]
     (select-keys m [:line :file :doc])))
 
-(mx/defn test-case-result
+(defn test-case-result
   "Creates a test case result map with keys :pass?, :source, and :thrown.
 
   pass? is true if the test case passed successfully, false otherwise.
@@ -51,9 +45,7 @@
 
   thrown is the exception (Throwable) thrown by a failing test case."
   ([type' source] (test-case-result type' source nil))
-  ([type'
-    source :- [:fn test-case?]
-    thrown :- [:maybe :lt/throwable]]
+  ([type' source thrown]
    (let [{:keys [file line doc]} (extract-file-line-doc source thrown)
          {:keys [actual expected message caught evaluated]} (ex-data thrown)
          thrown' (or (when (instance? Throwable caught) caught) thrown)]
@@ -75,14 +67,14 @@
 (defn identifier [result]
   (or (:doc result) "Anonymous test case"))
 
-(mx/defn try-test-case
+(defn try-test-case
   "Executes a test case function. Catches all Throwables. Returns a
    map with the following key-value pairs:
 
      :source - the input function
      :pass?  - true if the function ran without throwing
      :thrown - the Throwable instance if thrown"
-  [f :- [:fn test-case?]]
+  [f]
   (try (f)
     (test-case-result :pass f)
     (catch Throwable t